Move Student plan list to separate directory
